memory-benchmark/requirements.txt
# Core dependencies
mem0ai>=0.0.1
zep-cloud>=3.16.0

# Evaluation dependencies
anthropic>=0.18.0  # For LLM-based semantic evaluation (optional)

# Utilities
python-dotenv>=1.0.0
